+++ test/replay/incentive.go
@@ -6,8 +6,8 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// TODO(rafilx): after coostaking genesis, add test checking rewards portions are correctly taken from
-// incentive, coostaaking, distribution each block from the fee collector
+// TODO(rafilx): after costaking genesis, add test checking rewards portions are correctly taken from
+// incentive, costaking, distribution each block from the fee collector
 
 func (d *BabylonAppDriver) GenerateBlocksUntilLastProcessedBtcStkEventsHeightIs(untilBlock uint64) {
 	ictvK := d.App.IncentiveKeeper